stress-dirmany: cast ptr diff to size_t to clear up pedantic clang warning

+++ stress-dirmany.c
@@ -59,7 +59,7 @@ static void stress_dirmany_filename(
 		*ptr++ = '/';
 		(void)memset(ptr, 'x', filename_len);
 		ptr += filename_len;
-		(void)snprintf(ptr, sizeof(filename) + (ptr - filename), "%16.16" PRIx64, n);
+		(void)snprintf(ptr, sizeof(filename) + (size_t)(ptr - filename), "%16.16" PRIx64, n);
 	} else {
 		(void)snprintf(filename, filename_sz, "%16.16" PRIx64, n);
 	}
